End screen when completing level
When all enemies of all waves are defeated, a new scene is created with
the "choose your next skill" dialog (empty for now).
I've chosen to create a new scene instead of adding elements on the
current one to not have to bother about making sure the player can't
fire skills or interact with the world anymore.
We need to persist the state of the game between scenes though, but this
is something we will have to do to handle save games anyway